+++ sdk/storage/Azure.Storage.Blobs/src/PartitionedDownloader.cs
@@ -89,7 +89,9 @@ namespace Azure.Storage.Blobs
             }
             else
             {
-                _maxWorkerCount = Constants.Blob.Block.DefaultConcurrentTransfersCount;
+                _maxWorkerCount = CompatSwitches.UseLegacyDefaultConcurrency
+                    ? Constants.Blob.Block.LegacyDefaultConcurrentTransfersCount
+                    : Environment.ProcessorCount;
             }
 
             // Set _rangeSize